// landing/fragments/landing-price/src/content/content-without-installments/content-without-installments.interfaces.ts
export interface ContentWithoutInstallmentsProps {
  costPerMonth?: string
  currency?: string
  oldPrice?: string
  quantityVideoLessons?: string
  quantityMonths?: string
  backgroundRectangle?: string
  squareRotate?: number
  rectangleRotate?: number
  rectanglePositionX?: number
  rectanglePositionY?: number
}